Remove data-loaders from `FaciesInfo` and `FaciesCubeset`
Due to obscurity of geometry-labels correpondence loading process,
from now it must be explicitly done in the notebook to allow easy debug.

Instead of creating cubeset and loading geometry/labels, `FaciesInfo`
provides now a list of cubes files paths (for `SeismicCubeset.__init__)
and dictionaries of cube-labels correpondence for labels to load
for (`SeismicCubeset.create_labels`).