Make `SesmicSampler` plots better, infer ncols/nrows in `MatplotlibPlotter`
Place all sampler plots on a single figure, instead of plotting them separately.
Plot legend in sampler plots on a separate axis instead of putting it above the plot.
Make a separate procedure for inferring number of figure cols, rows if not provided.
Take `xlim` and `ylim` params in account when infering default figure size.